first and last episode, season
